+++ X10D.Unity/src/DebugUtility/DebugUtility.Circle.cs
@@ -6,7 +6,7 @@ using Quaternion = System.Numerics.Quaternion;
 
 namespace X10D.Unity;
 
-public static partial class DebugEx
+public static partial class DebugUtility
 {
     /// <summary>
     ///     Draws a circle with the specified color.