+++ X10D.Unity/src/DebugUtility/DebugUtility.Ray.cs
@@ -2,7 +2,7 @@
 
 namespace X10D.Unity;
 
-public static partial class DebugEx
+public static partial class DebugUtility
 {
     /// <summary>
     ///     Draws a ray.